requirements.txt
aiofiles
aiohttp==3.11.12
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
attrs==25.4.0
audioop-lts; python_version >= "3.13"
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blis==1.3.3
build==1.4.0
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.5.2
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.5
cymem==2.0.13
dataclasses-json==0.6.7
datasets==4.6.0
deepeval
Deprecated==1.3.1
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
ecdsa==0.19.1
emoji==2.15.0
et_xmlfile==2.0.0
execnet==2.1.2
faiss-cpu==1.13.2
fastapi==0.134.0
filelock==3.24.3
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-vision==3.12.1
googleapis-common-protos==1.72.0
greenlet==3.3.2
groq==0.37.1
gunicorn==23.0.0
grpcio==1.78.0
grpcio-status==1.78.0
gTTS==2.2.4
h11==0.16.0
hf-xet==1.3.2
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
installer==0.7.0
instructor
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain>=0.3.0
langchain-classic>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
langdetect==1.0.9
langfuse==3.14.5
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.3.45
limits==5.8.0
llvmlite==0.46.0
lxml==6.0.2
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
messagebird==2.2.0
mmh3==5.2.0
mpmath==1.3.0
msoffcrypto-tool==6.0.0
multidict==6.7.1
multiprocess==0.70.18
murmurhash==1.0.15
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
numba==0.64.0
numpy
oauthlib==3.3.1
olefile==0.47
onnxruntime==1.24.2
openai
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==24.2
pandas==3.0.1
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20260107
pi_heif==1.3.0
pikepdf==10.3.0
pillow==12.1.1
pluggy==1.6.0
portalocker==3.2.0
posthog==5.4.0
preshed==3.0.12
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
psycopg2-binary==2.9.11
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydub==0.25.1
pyfiglet==1.0.4
Pygments==2.19.2
PyJWT==2.11.0
pyodbc==5.3.0
pypdf==6.7.4
pypdfium2==5.5.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-repeat==0.9.4
pytest-rerunfailures==16.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-iso639==2026.1.31
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.22
python-oxmsg==0.0.2
python-pptx==1.0.2
PyYAML==6.0.3
ragas==0.4.3
rank-bm25==0.2.2
RapidFuzz==3.14.3
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
scikit-network==0.33.5
scipy==1.17.1
sentry-sdk==2.53.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
smart_open==7.5.1
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.5
SQLAlchemy==2.0.47
srsly==2.5.2
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.52.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
thinc==8.3.10
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.21.5
unstructured-client==0.42.10
unstructured.pytesseract==0.3.15
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.30.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wheel==0.46.3
wrapt==1.17.3
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.23.0
